move antlr dependency from root module to parser engine module
