+++ shardingsphere-sql-parser/shardingsphere-sql-parser-engine/pom.xml
@@ -47,5 +47,9 @@
             <groupId>org.apache.commons</groupId>
             <artifactId>commons-collections4</artifactId>
         </dependency>
+        <dependency>
+            <groupId>org.antlr</groupId>
+            <artifactId>antlr4-runtime</artifactId>
+        </dependency>
     </dependencies>
 </project>
