+++ shardingsphere-sql-parser/shardingsphere-sql-parser-engine/src/test/java/org/apache/shardingsphere/sql/parser/core/visitor/VisitorRuleTest.java
@@ -15,9 +15,23 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.sql.parser.sql.constant;
+package org.apache.shardingsphere.sql.parser.core.visitor;
 
 import org.antlr.v4.runtime.tree.ParseTree;
+import org.junit.Test;
 
-public interface SelectContext extends ParseTree {
+import static org.hamcrest.CoreMatchers.is;
+import static org.junit.Assert.assertThat;
+
+public final class VisitorRuleTest {
+    
+    @Test
+    public void assertValueOfParseTreeClassSuccess() {
+        assertThat(VisitorRule.valueOf(SelectContext.class), is(VisitorRule.SELECT));
+    }
+    
+    @Test(expected = IllegalArgumentException.class)
+    public void assertValueOfParseTreeClassFailure() {
+        VisitorRule.valueOf(ParseTree.class);
+    }
 }
