+++ shardingsphere-sql-parser/shardingsphere-sql-parser-postgresql/src/main/java/org/apache/shardingsphere/sql/parser/PostgreSQLParser.java
@@ -21,7 +21,7 @@ import org.antlr.v4.runtime.TokenStream;
 import org.apache.shardingsphere.sql.parser.api.ASTNode;
 import org.apache.shardingsphere.sql.parser.api.parser.SQLParser;
 import org.apache.shardingsphere.sql.parser.autogen.PostgreSQLStatementParser;
-import org.apache.shardingsphere.sql.parser.core.DefaultASTNode;
+import org.apache.shardingsphere.sql.parser.core.ParseASTNode;
 
 /**
  * SQL parser for PostgreSQL.
@@ -34,6 +34,6 @@ public final class PostgreSQLParser extends PostgreSQLStatementParser implements
     
     @Override
     public ASTNode parse() {
-        return new DefaultASTNode(execute());
+        return new ParseASTNode(execute());
     }
 }
