+++ shardingsphere-sql-parser/shardingsphere-sql-parser-sql92/src/main/java/org/apache/shardingsphere/sql/parser/SQL92Parser.java
@@ -21,7 +21,7 @@ import org.antlr.v4.runtime.TokenStream;
 import org.apache.shardingsphere.sql.parser.api.ASTNode;
 import org.apache.shardingsphere.sql.parser.api.parser.SQLParser;
 import org.apache.shardingsphere.sql.parser.autogen.SQL92StatementParser;
-import org.apache.shardingsphere.sql.parser.core.DefaultASTNode;
+import org.apache.shardingsphere.sql.parser.core.ParseASTNode;
 
 /**
  * SQL parser for SQL92.
@@ -34,6 +34,6 @@ public final class SQL92Parser extends SQL92StatementParser implements SQLParser
     
     @Override
     public ASTNode parse() {
-        return new DefaultASTNode(execute());
+        return new ParseASTNode(execute());
     }
 }
